fix: @ApiPropertyの補完を効かせるためにDTOを切り出す
