+++ backend/src/app/participant/dto/participant.dto.ts
@@ -1,0 +1,11 @@
+export class ParticipantDTO {
+    public readonly id: string
+    public readonly name: string
+    public readonly enrollementStatus: string
+    public constructor(props: { id: string; name: string; enrollementStatus: string }) {
+        const { id, name, enrollementStatus } = props
+        this.id = id
+        this.name = name
+        this.enrollementStatus = enrollementStatus
+    }
+}
