+++ backend/src/app/participant/query-service-interface/participant-repository.ts
@@ -1,14 +1,4 @@
-export class ParticipantDTO {
-  public readonly id: string
-  public readonly name: string
-  public readonly enrollementStatus: string
-  public constructor(props: { id: string; name: string; enrollementStatus: string }) {
-    const { id, name, enrollementStatus } = props
-    this.id = id
-    this.name = name
-    this.enrollementStatus = enrollementStatus
-  }
-}
+import { ParticipantDTO } from "src/app/participant/dto/participant.dto"
 
 export interface ParticipantRepository {
   getAll(): Promise<ParticipantDTO[]>
